requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-pymongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.6.0
Werkzeug==3.1.3
python-dotenv==1.0.1
